+++ CHANGELOG.md
@@ -5,7 +5,8 @@ of Concierge. Please check the Wiki entry on the release process to understand
 how this file is formatted and how the process works.
 
 ## Unreleased
--
+### Added
+- Avantio: handle booking not available case in common manner
 
 ## [1.3.3] - 2016-11-23
 ### Changed